Move controlling tests into oban_test
